+++ perf/tests/test_perf_cli.py
@@ -140,7 +140,7 @@ class TestPerfCLI(BaseTestCase, unittest.TestCase):
         stdout = self.compare('compare_to', ref_result, changed_result, '-v')
 
         expected = ('Median +- std dev: [ref] 1.50 sec +- 0.50 sec '
-                    '-> [changed] 2.00 sec +- 0.50 sec: 1.33x slower\n'
+                    '-> [changed] 2.00 sec +- 0.50 sec: 1.33x slower (+33%)\n'
                     'Not significant!')
         self.assertEqual(stdout.rstrip(),
                          expected)
@@ -166,7 +166,7 @@ class TestPerfCLI(BaseTestCase, unittest.TestCase):
         stdout = self.compare('compare', ref_result, changed_result, '-v')
 
         expected = ('Median +- std dev: [ref] 1.50 sec +- 0.50 sec '
-                    '-> [changed] 2.00 sec +- 0.50 sec: 1.33x slower\n'
+                    '-> [changed] 2.00 sec +- 0.50 sec: 1.33x slower (+33%)\n'
                     'Not significant!')
         self.assertEqual(stdout.rstrip(),
                          expected)